정환님과 함께한 깃허브 버그 고치기
This reverts commit d09cfdc9f2a49ff7fbecec29c15d925f23c17cee, reversing
changes made to a3d6a71620803fa71aea8d03cbfeb11a25fe5028.

+++ Source/RuinLooters/UI/HPWidget.cpp
@@ -4,14 +4,14 @@
 #include "UI/HPWidget.h"
 #include "Components/ProgressBar.h"
 
-void UHPWidget::CalculateHp(int32 NewCurrentHp, int32 NewMaxHp)
+void UHPWidget::CalculateHp(float NewCurrentHp, float NewMaxHp)
 {
     if (!HpBar) return;
 
     // SetPercent()가 float 타입을 받기 때문에 타입변환
-    float Ratio = (float)NewCurrentHp / NewMaxHp;
+    float Ratio = NewCurrentHp / NewMaxHp;
     HpBar->SetPercent(Ratio);
-    UE_LOG(LogTemp, Warning, TEXT("%d / %d"), NewCurrentHp, NewMaxHp);
+    UE_LOG(LogTemp, Warning, TEXT("%.1f / %.1f"), NewCurrentHp, NewMaxHp);
 }
 
 // 적이 죽 위젯 삭제
