+++ Source/RuinLooters/UI/RLPlayerUI.cpp
@@ -36,11 +36,11 @@ void URLPlayerUI::NativeTick(const FGeometry& MyGeometry, float InDeltaTime)
 	}
 }
 
-void URLPlayerUI::PlayerCalculateHp(int32 NewCurrentHp, int32 NewMaxHp)
+void URLPlayerUI::PlayerCalculateHp(float NewCurrentHp, float NewMaxHp)
 {
 
 	if (!PlayerHpBar) return;
-	float Ratio = (float)NewCurrentHp / NewMaxHp;
+	float Ratio = NewCurrentHp / NewMaxHp;
 	PlayerHpBar->SetPercent(Ratio);
 }
